+++ cnpj-receita-federal-brasil/core/br/gov/economia/receita/imp/ILayout.java
@@ -3,8 +3,8 @@ package br.gov.economia.receita.imp;
 import br.gov.economia.receita.IRegisterVisitor;
 
 interface ILayout {
-  public VisitResult begin(IRegisterVisitor visitor, int row);
-  public VisitResult visit(IRegisterVisitor visitor, int row, String line);
+  public VisitResult begin(IRegisterVisitor visitor, long row);
+  public VisitResult visit(IRegisterVisitor visitor, long row, String line);
   public VisitResult end(IRegisterVisitor visitor);
 }
 
@@ -12,12 +12,12 @@ enum SkipLayout implements ILayout {
   INSTANCE;
   
   @Override
-  public VisitResult visit(IRegisterVisitor visitor, int row, String line) {
+  public VisitResult visit(IRegisterVisitor visitor, long row, String line) {
     return VisitResult.SKIP;
   }
 
   @Override
-  public VisitResult begin(IRegisterVisitor visitor, int row) {
+  public VisitResult begin(IRegisterVisitor visitor, long row) {
     return VisitResult.SKIP;
   }
