+++ cnpj-receita-federal-brasil/visitor/br/gov/economia/receita/imp/SocioVisitor.java
@@ -6,12 +6,12 @@ import br.gov.economia.receita.IRegisterVisitor;
 public abstract class SocioVisitor implements IRegisterVisitor{
 
   @Override
-  public VisitResult beginHeader(int row) {
+  public VisitResult beginHeader(long row) {
     return VisitResult.SKIP;
   }
 
   @Override
-  public VisitResult fieldHeader(int row, IField field) {
+  public VisitResult fieldHeader(long row, IField field) {
     return VisitResult.SKIP;
   }
 
@@ -21,12 +21,12 @@ public abstract class SocioVisitor implements IRegisterVisitor{
   }
 
   @Override
-  public VisitResult beginCnae(int row) {
+  public VisitResult beginCnae(long row) {
     return VisitResult.SKIP;
   }
 
   @Override
-  public VisitResult fieldCnae(int row, IField field) {
+  public VisitResult fieldCnae(long row, IField field) {
     return VisitResult.SKIP;
   }
 
@@ -36,12 +36,12 @@ public abstract class SocioVisitor implements IRegisterVisitor{
   }
 
   @Override
-  public VisitResult beginTrailler(int row) {
+  public VisitResult beginTrailler(long row) {
     return VisitResult.SKIP;
   }
 
   @Override
-  public VisitResult fieldTrailler(int row, IField field) {
+  public VisitResult fieldTrailler(long row, IField field) {
     return VisitResult.SKIP;
   }
 
@@ -51,12 +51,12 @@ public abstract class SocioVisitor implements IRegisterVisitor{
   }
 
   @Override
-  public VisitResult beginEmpresa(int row) {
+  public VisitResult beginEmpresa(long row) {
     return VisitResult.SKIP;
   }
 
   @Override
-  public VisitResult fieldEmpresa(int row, IField field) {
+  public VisitResult fieldEmpresa(long row, IField field) {
     return VisitResult.SKIP;
   }
 
@@ -66,7 +66,7 @@ public abstract class SocioVisitor implements IRegisterVisitor{
   }
 
   @Override
-  public VisitResult handleError(int lineNumber, Throwable e) {
+  public VisitResult handleError(long lineNumber, Throwable e) {
     e.printStackTrace(System.err);
     return VisitResult.TERMINATE;
   }
